refactor: schema 파일 crdt 폴더 내로 이동
- crdt에서 일어나는 건 crdt폴더 내에서만 일어나도록 feature화

#122
