+++ server/src/crdt/crdt.module.ts
@@ -1,7 +1,7 @@
 import { Module } from "@nestjs/common";
 import { CrdtService } from "./crdt.service";
 import { MongooseModule } from "@nestjs/mongoose";
-import { Doc, DocumentSchema } from "../schemas/document.schema";
+import { Doc, DocumentSchema } from "./schemas/document.schema";
 import { CrdtGateway } from "./crdt.gateway";
 
 @Module({
